Saving _and_ removing. w00t!
So I'm not happy with the implementation, I think I was over-thinking it _a lot_.

But it's done, and test coverage only dipped a bit due to not having an `else` (urgh).

The `ButtonAdd` component. I was trying to be too clever in making a reusuable button component, but actually, for this small thing I was only needing it for one line template.

It didn't make sense to have a one line template by itself, so I deleted it and merged it with the `VideoCard` component.

PR Feedback

Included comments
Removed unused test spec

Updating e2e tests

Now tests if the button is actually clicked and button text is updated accordingly.